A few change ofr the script
Hi,
I first contact Bo to pull and give my vision of your script. He redirected me to you.
So I pull my change to GIT and you will do what you want to do with it.

The big change in my way is that the script is not Fedora only.
It permit to use the actual displays xrandr configuration for xinitrc
and to set permanently the display to Nvidia on each start.
It also use complete xorg conf for each card and optional config file for custon nvidia install
or for other distro.

I thought to create an alternate project, maybe a better way to not pollute yours.

Tell me your thoughts.

My only big issue was I didn't nicely manage the white space for the xrandr screen list.
If your have a better way, be welcome (line 81 to 84)

Have a nice day.